backend/requirements.txt
# FastAPI 核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# 环境变量
python-dotenv==1.0.0

# 工具
PyYAML==6.0.1
pandas==2.1.3
openpyxl==3.1.2

